services/ingestion/requirements.txt
fastapi==0.115.6
uvicorn[standard]==0.34.0
sqlalchemy==2.0.37
psycopg2-binary==2.9.10
redis==5.2.1
httpx==0.28.1
pydantic==2.10.5
pydantic-settings==2.7.1
structlog==25.1.0

# OpenTelemetry
opentelemetry-api==1.29.0
opentelemetry-sdk==1.29.0
opentelemetry-exporter-otlp-proto-grpc==1.29.0
opentelemetry-instrumentation-fastapi==0.50b0
opentelemetry-instrumentation-sqlalchemy==0.50b0
opentelemetry-instrumentation-redis==0.50b0
opentelemetry-instrumentation-httpx==0.50b0
